requirements.txt
asgiref==3.2.3
backcall==0.1.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0
django-annoying==0.10.6
django-autocomplete-light==3.8.1
django-crontab==0.7.1
django-import-export==2.0
django-material-admin==1.6.11
django-more-admin-filters==1.3
django-rest-framework==0.1.0
django-fsm==2.7.1
django-simple-history==2.8.0
#django-suit==2.0a1
django-daterangefilter==1.0.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
ipython==7.11.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.2
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
python-dotenv==0.15.0
pytz==2019.3
PyYAML==5.2
requests==2.25.1
urllib3==1.26.3
https://github.com/darklow/django-suit/tarball/v2